rework template system, update helix config
rather than having a monolithic template and a rigid classification
system, just have a template for each editor and template only the
matcher portion. this is much more flexible. it's really, really hard to
have a single template with polar opposite precedence rules.

in fact, if we really want to go down to a single template, we could
just pass the result through something that reverses the order of all
the rules.